gas_booking/requirements.txt
anyascii==0.2.0
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.4 (install)
django-allauth==0.44.0  (install)
django-classy-tags==2.0.0
django-cms==3.8.0  (install)
django-cogwheels==0.3
django-countries==7.2.1  (install)
django-filter==2.4.0
django-formtools==2.3
django-modelcluster==5.1
django-sekizai==2.0.0
django-taggit==1.4.0
django-treebeard==4.5.1
djangocms-admin-style==2.0.2
djangorestframework==3.12.4
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
html5lib==1.1
idna==2.10
l18n==2020.6.1
mysqlclient==2.0.3  (install)
oauthlib==3.1.1
openpyxl==3.0.7
Pillow==8.2.0
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
telepath==0.2
urllib3==1.26.5
wagtail==2.13.2
wagtailmenus==3.0.2 
webencodings==0.5.1
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.4.3
xlwt==1.3.0
